SignalR聊天功能v1.0版。Hub端功能(聊天紀錄能儲存到資料庫、能夠從資料庫獲取歷史聊天紀錄)，Client端使用Vue實作與Hub的連線。

+++ SuFood專案MVC/SuFood/Controllers/CheckOutController.cs
@@ -56,12 +56,12 @@ namespace SuFood.Controllers
 			Orders od = new Orders
 			{
 				OrdersId = vmParameters.OrdersId,
-				SubTotal = vmParameters.SubTotal,
+				SubTotal = (int)vmParameters.SubTotal,
 				SetOrdersDatetime = DateTime.Now.AddMilliseconds(-DateTime.Now.Millisecond),
 				ShipAddress = vmParameters.ShipAddress,
 				CouponId = vmParameters.CouponId,
 				OrdersDetails = vmParameters.OrdersDetails,
-				AccountId = vmParameters.AccountId
+				AccountId = (int)vmParameters.AccountId
 			};
 
 			_context.Orders.Add(od);
